+++ src/main/java/com/example/springboot/app/repository/AirportRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface AirportRepository extends JpaRepository<Airport, Long> {
+public interface AirportRepository extends JpaRepository<Airport, Integer> {
     // Con @Repository le indico los metodos principales select, create, update, delete
 }
